+++ src/main.cpp
@@ -1,11 +1,17 @@
-#include <Arduino.h>
+// SPDX-License-Identifier: GPL-2.0-or-later WITH GCC-exception-2.0
 
+#include <Arduino.h>
+#include <EspBacnet.h>
 
 #ifndef PIO_UNIT_TESTING
-void setup(){
+void setup() {
+  BacnetClient client;
+  BacnetServer server;
 
+  (void)client;
+  (void)server;
 }
-void loop(){
 
+void loop() {
 }
 #endif